requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
python-dotenv==1.0.0
PyMuPDF==1.23.8
python-docx==1.1.0
emoji==2.10.0
Pillow==10.2.0
transformers==4.36.0
tiktoken==0.5.2

